aiservice/pyproject.toml
[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[project]
name = "aiservice"
version = "0.1.0"
description = "Azure OpenAI authentication wrapper"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.0",
    "azure-identity>=1.15",
    "httpx>=0.27",
]

[tool.pdm.build]
package-dir = ".."
includes = ["aiservice/"]
